relocate profile page jsx to pages dir
